requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.2.1
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cssselect2==0.4.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-bootstrap4==2.2.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
packaging==20.7
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.3
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
reportlab==3.5.55
requests==2.25.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.56.0
terminado==0.9.1
testpath==0.4.4
tinycss2==1.1.0
toml==0.10.1
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
virtualenv==20.4.2
wcwidth==0.2.5
WeasyPrint==52.1
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.5